requirements.txt
alias-free-torch>=0.0.6
antlr4-python3-runtime>=4.9.3
cffi>=2.0.0
diffusers>=0.37.1
einops>=0.8.2
einops-exts>=0.0.4
einx>=0.4.3
frozendict>=2.4.7
julius>=0.2.7
lameenc>=1.8.2
omegaconf>=2.3.0
openunmix>=1.3.0
pycparser>=3.0
safetensors>=0.7.0
soundfile>=0.13.1
transformers>=5.5.0
tqdm>=4.67.3
vector-quantize-pytorch>=1.28.1
